I add a lot of changes like seed fake data and improving controllers...

+++ controller/articleController.js
@@ -6,9 +6,14 @@ const getAllArticles = async (req, res) => {
     const articles = await prisma.article.findMany({
       include: {
         author: true,
+        comments: true,
+        categories: true,
       },
     });
-    res.status(200).json(articles);
+    res.status(200).json({
+      message: "All articles retrieved successfully",
+      articles,
+    });
   } catch (error) {
     res.status(500).json({ error: error.message });
   }
@@ -18,15 +23,34 @@ const getAllArticles = async (req, res) => {
 const getOneArticle = async (req, res) => {
   try {
     const { id } = req.params;
+
+    //Check if article exists
+    const articleExists = await prisma.article.findUnique({
+      where: {
+        id: parseInt(id),
+      },
+    });
+
+    if (!articleExists) {
+      return res.status(404).json({
+        error: "Article not found",
+      });
+    }
+    //Get article
     const article = await prisma.article.findUnique({
       where: {
         id: parseInt(id),
       },
       include: {
         author: true,
+        comments: true,
+        categories: true,
       },
     });
-    res.status(200).json(article);
+    res.status(200).json({
+      message: "Article retrieved successfully",
+      article,
+    });
   } catch (error) {
     res.status(500).json({ error: error.message });
   }
@@ -36,6 +60,7 @@ const getOneArticle = async (req, res) => {
 const createArticle = async (req, res) => {
   try {
     const { title, content, image, authorId } = req.body;
+
     const newArticle = await prisma.article.create({
       data: {
         title,
@@ -43,11 +68,11 @@ const createArticle = async (req, res) => {
         image,
         author: { connect: { id: authorId } },
       },
-      include: {
-        author: true,
-      },
     });
-    res.status(201).json(newArticle);
+    res.status(201).json({
+      message: "Article created successfully",
+      newArticle,
+    });
   } catch (error) {
     res.status(500).json({ error: error.message });
   }
@@ -57,7 +82,22 @@ const createArticle = async (req, res) => {
 const updateArticle = async (req, res) => {
   try {
     const { id } = req.params;
-    const { title, content, image, authorId } = req.body;
+    const { title, content, image } = req.body;
+
+    //Check if article exists
+    const articleExists = await prisma.article.findUnique({
+      where: {
+        id: parseInt(id),
+      },
+    });
+
+    if (!articleExists) {
+      return res.status(404).json({
+        error: "Article you want to update not found",
+      });
+    }
+
+    //Update article
     const updateArticle = await prisma.article.update({
       where: {
         id: parseInt(id),
@@ -70,7 +110,10 @@ const updateArticle = async (req, res) => {
       },
     });
 
-    res.status(200).json(updateArticle);
+    res.status(200).json({
+      message: "Article updated successfully",
+      updateArticle,
+    });
   } catch (error) {
     res.status(500).json({ error: error.message });
   }
@@ -80,13 +123,37 @@ const updateArticle = async (req, res) => {
 const deleteArticle = async (req, res) => {
   try {
     const { id } = req.params;
-    const deleteArticle = await prisma.article.delete({
+
+    //Check if article exists
+    const articleExists = await prisma.article.findUnique({
       where: {
         id: parseInt(id),
       },
     });
 
-    res.status(200).json(deleteArticle);
+    if (!articleExists) {
+      return res.status(404).json({
+        error: "Article you want to delete not found",
+      });
+    }
+
+    //Delete associated comments
+    await prisma.comment.deleteMany({
+      where: {
+        articleId: parseInt(id),
+      },
+    });
+
+    //Delete article
+    await prisma.article.delete({
+      where: {
+        id: parseInt(id),
+      },
+    });
+
+    res.status(204).json({
+      message: "Article deleted successfully",
+    });
   } catch (error) {
     res.status(500).json({ error: error.message });
   }
