+++ controller/commentController.js
@@ -4,6 +4,30 @@ const prisma = require("../server");
 const createComment = async (req, res) => {
   try {
     const { content, authorId, articleId } = req.body;
+
+    //Check if the article exists
+    const article = await prisma.article.findUnique({
+      where: {
+        id: articleId,
+      },
+    });
+
+    if (!article) {
+      return res.status(404).json({ message: "Article not found" });
+    }
+
+    //Check if the author exists
+    const author = await prisma.user.findUnique({
+      where: {
+        id: authorId,
+      },
+    });
+
+    if (!author) {
+      return res.status(404).json({ message: "Author not found" });
+    }
+
+    //Create a new comment
     const newComment = await prisma.comment.create({
       data: {
         content,
@@ -16,7 +40,10 @@ const createComment = async (req, res) => {
       },
     });
 
-    res.status(201).json(newComment);
+    res.status(201).json({
+      message: "Comment created successfully",
+      newComment,
+    });
   } catch (error) {
     res.status(500).json({ message: error });
   }
@@ -26,16 +53,62 @@ const createComment = async (req, res) => {
 const getAllComments = async (req, res) => {
   try {
     const { id } = req.params;
+
+    //Check if the article exists
+    const article = await prisma.article.findUnique({
+      where: {
+        id: parseInt(id),
+      },
+    });
+
+    if (!article) {
+      return res.status(404).json({ message: "Article not found" });
+    }
+
+    //Retrieve all comments for a specific article
     const comments = await prisma.comment.findMany({
       where: {
         articleId: parseInt(id),
       },
       include: {
         article: true,
+        author: true,
       },
     });
 
     res.status(200).json(comments);
+  } catch (error) {
+    res.status(500).json({ message: error });
+  }
+};
+
+//Get One Comment
+const getOneComment = async (req, res) => {
+  try {
+    const { id } = req.params;
+
+    //Check if the comment exists
+    const commentExist = await prisma.comment.findUnique({
+      where: {
+        id: parseInt(id),
+      },
+    });
+
+    if (!commentExist) {
+      return res.status(404).json({ message: "Comment not found" });
+    }
+
+    //Retrieve a comment
+    const comment = await prisma.comment.findUnique({
+      where: {
+        id: parseInt(id),
+      },
+    });
+
+    res.status(200).json({
+      message: "Comment retrieved successfully",
+      comment,
+    });
   } catch (error) {
     res.status(500).json({ message: error });
   }
@@ -47,6 +120,30 @@ const updateComment = async (req, res) => {
     const { id } = req.params;
     const { content } = req.body;
 
+    //Check if the comment exists
+    const comment = await prisma.comment.findUnique({
+      where: {
+        id: parseInt(id),
+      },
+    });
+
+    if (!comment) {
+      return res.status(404).json({ message: "Comment not found" });
+    }
+
+    //Check if the content is empty
+    if (!content) {
+      return res.status(400).json({ message: "Content is required" });
+    }
+
+    //Check if the content is the same as the previous one
+    if (content === comment.content) {
+      return res.status(400).json({
+        message: "The content is the same as the previous one",
+      });
+    }
+
+    //Update the comment
     const updateComment = await prisma.comment.update({
       where: {
         id: parseInt(id),
@@ -65,13 +162,29 @@ const updateComment = async (req, res) => {
 const deleteComment = async (req, res) => {
   try {
     const { id } = req.params;
+
+    //Check if the comment exists
+    const commentExist = await prisma.comment.findUnique({
+      where: {
+        id: parseInt(id),
+      },
+    });
+
+    if (!commentExist) {
+      return res.status(404).json({ message: "Comment not found" });
+    }
+
+    //Delete a comment
     const deleteComment = await prisma.comment.delete({
       where: {
         id: parseInt(id),
       },
     });
 
-    res.status(200).json(deleteComment);
+    res.status(200).json({
+      message: "Comment deleted successfully",
+      deleteComment,
+    });
   } catch (error) {
     res.status(500).json({ message: error });
   }
@@ -80,6 +193,7 @@ const deleteComment = async (req, res) => {
 module.exports = {
   createComment,
   getAllComments,
+  getOneComment,
   updateComment,
   deleteComment,
 };
